+++ pages/signin.tsx
@@ -1,4 +1,4 @@
-import Input from '@/components/common/Input';
+import Input from '@/components/common/Form';
 
 const SignIn = () => {
   return <Input />;
